Tax moved to teh new package, tax calculation implemented in the static method Tax.evaluateTax(RequestItem item)
