+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/invoicing/BookKeeper.java
@@ -20,6 +20,7 @@ import java.util.List;
 
 import pl.com.bottega.ecommerce.canonicalmodel.publishedlanguage.ClientData;
 import pl.com.bottega.ecommerce.canonicalmodel.publishedlanguage.Id;
+import pl.com.bottega.ecommerce.sales.domain.tax.Tax;
 import pl.com.bottega.ecommerce.sharedkernel.Money;
 
 public class BookKeeper {
@@ -28,33 +29,10 @@ public class BookKeeper {
 		Invoice invoice = InvoiceFactory.createInvoice(Id.generate(), request.getClient());
 
 		for (RequestItem item : request.getItems()) {
+			
 			Money net = item.getTotalCost();
-			BigDecimal ratio = null;
-			String desc = null;
-			
-			switch (item.getProductData().getType()) {
-			case DRUG:
-				ratio = BigDecimal.valueOf(0.05);
-				desc = "5% (D)";
-				break;
-			case FOOD:
-				ratio = BigDecimal.valueOf(0.07);
-				desc = "7% (F)";
-				break;
-			case STANDARD:
-				ratio = BigDecimal.valueOf(0.23);
-				desc = "23%";
-				break;
-				
-			default:
-				throw new IllegalArgumentException(item.getProductData().getType() + " not handled");
-			}
-					
-			Money taxValue = net.multiplyBy(ratio);
-			
-			Tax tax = new Tax(taxValue, desc);
-			
-
+			Tax tax = Tax.evaluateTax(item);
+	
 			InvoiceLine invoiceLine = new InvoiceLine(item.getProductData(),
 					item.getQuantity(), net, tax);
 			invoice.addItem(invoiceLine);
