+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/invoicing/RequestItem.java
@@ -18,7 +18,7 @@ package pl.com.bottega.ecommerce.sales.domain.invoicing;
 import pl.com.bottega.ecommerce.sales.domain.productscatalog.ProductData;
 import pl.com.bottega.ecommerce.sharedkernel.Money;
 
-class RequestItem {
+public class RequestItem {
 
 	private ProductData productData;
