+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/tax/Tax.java
@@ -13,9 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package pl.com.bottega.ecommerce.sales.domain.invoicing;
+package pl.com.bottega.ecommerce.sales.domain.tax;
 
 import pl.com.bottega.ecommerce.sharedkernel.Money;
+import java.math.BigDecimal;
+import pl.com.bottega.ecommerce.sales.domain.invoicing.RequestItem;
 
 public class Tax {
 
@@ -35,6 +37,32 @@ public class Tax {
 
 	public String getDescription() {
 		return description;
+	}
+	
+	public static Tax evaluateTax(RequestItem item) {
+		Money net = item.getTotalCost();
+		BigDecimal ratio = null;
+		String desc = null;
+		
+		switch (item.getProductData().getType()) {
+		case DRUG:
+			ratio = BigDecimal.valueOf(0.05);
+			desc = "5% (D)";
+			break;
+		case FOOD:
+			ratio = BigDecimal.valueOf(0.07);
+			desc = "7% (F)";
+			break;
+		case STANDARD:
+			ratio = BigDecimal.valueOf(0.23);
+			desc = "23%";
+			break;
+		default:
+			throw new IllegalArgumentException(item.getProductData().getType() + " not handled");	
+		}					
+		Money taxValue = net.multiplyBy(ratio);
+		
+		return new Tax(taxValue, desc);
 	}
 
 }
